# AST: Remodel `MemberTypeRepr` to be recursive