… lib/AST/ASTPrinter.cpp
… lib/AST/Attr.cpp
… lib/AST/Decl.cpp
… lib/AST/Expr.cpp
… lib/AST/NameLookup.cpp
… lib/AST/TypeCheckRequests.cpp
… lib/AST/TypeDeclFinder.cpp
… lib/ASTGen/Sources/ASTGen/Types.swift
… lib/IDE/Formatting.cpp
… lib/IDE/SourceEntityWalker.cpp
… lib/IDE/SyntaxModel.cpp
… lib/Migrator/APIDiffMigratorPass.cpp
… lib/Parse/ParseType.cpp
… lib/Sema/AssociatedTypeInference.cpp
… lib/Sema/CSDiagnostics.cpp
… lib/Sema/MiscDiagnostics.cpp
… lib/Sema/PreCheckExpr.cpp
… lib/Sema/TypeAccessScopeChecker.h
… lib/Sema/TypeCheckAccess.cpp
… lib/Sema/TypeCheckAvailability.cpp
… lib/Sema/TypeCheckDeclPrimary.cpp
… lib/Sema/TypeCheckPattern.cpp
… lib/Sema/TypeCheckType.cpp
… lib/Sema/TypeCheckType.h
… lib/Sema/TypeChecker.cpp
… test/Frontend/dump-parse.swift
… test/decl/protocol/req/assoc_type_inference_cycle.swift
… test/type/explicit_existential.swift
… test/type/explicit_existential_swift6.swift
… unittests/AST/ASTWalkerTests.cpp