A public interface that forces clients to interpret and traverse `MemberTypeRepr` nodes recursively such that every qualified type identifier is a `MemberTypeRepr` pointing to its qualifier:
* Is a more natural and bug-proof solution because each subsequent dot-separated identifier is not self-standing as in a tuple element. In the previous model, all type identifiers are standalone `IdentTypeRepr` nodes, and the only way to tell whether one is qualified is to look at the parent node, which is normally not a straightforward task with our AST. Knowledge of whether a type identifier is qualified is important in many of our analyses, but the absence of a general typed or other visible distinction between qualified and unqualified type identifiers is not helpful in guiding contributors to this realization.
* Grants more out-of-the-box flexibility to `ASTWalker` clients. `IdentTypeRepr` nodes are now unqualified by definition, and since each next dot-separated components is now itself a `MemberTypeRepr` and a parent of the previous one (vs. all components being direct descendants of a single `MemberTypeRepr`), children may be skipped halfway through a `MemberTypeRepr`.
* Will simplify modeling `.Type` and `.Protocol` types as `MemberTypeRepr`.