@AnthonyLatsis 2023-12-15
@swift-ci please smoke test macOS