@AnthonyLatsis 2023-12-16
@swift-ci please smoke test compiler performance