@AnthonyLatsis 2023-12-20
@swift-ci please smoke test macOS