@AnthonyLatsis 2023-12-20
@swift-ci please test compiler performance