@AnthonyLatsis 2023-12-21
# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-246b9f6963c2f40e)
    - [debug-batch brief](#debug-batch-brief-246b9f6963c2f40e)
    - [debug-batch detailed](#debug-batch-detailed-246b9f6963c2f40e)
- [Release](#release-246b9f6963c2f40e)
    - [release brief](#release-brief-246b9f6963c2f40e)
    - [release detailed](#release-detailed-246b9f6963c2f40e)



<a name="debug-batch-246b9f6963c2f40e">

# Debug-batch

<a name="debug-batch-brief-246b9f6963c2f40e">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 144,966,420,101,645 | 145,119,598,291,718 | 153,178,190,073 | 0.11%
LLVM.NumLLVMBytesOutput | 2,655,829,332 | 2,655,894,712 | 65,380 | 0.0%
</details>

<a name="debug-batch-detailed-246b9f6963c2f40e">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumPrecedenceGroups | 140 | 141 | 1 | 1.07% :no_entry:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (341)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 4,392,980 | 4,393,274 | 294 | 0.01%
AST.ImportSetCacheMiss | 804,563 | 804,630 | 67 | 0.01%
AST.ImportSetFoldHit | 405,177 | 405,232 | 55 | 0.01%
AST.ImportSetFoldMiss | 399,386 | 399,398 | 12 | 0.0%
AST.ModuleShadowCacheHit | 72,887 | 73,299 | 412 | 0.57%
AST.ModuleShadowCacheMiss | 5,359 | 5,377 | 18 | 0.34%
AST.ModuleVisibilityCacheHit | 93,136,473 | 93,153,183 | 16,710 | 0.02%
AST.ModuleVisibilityCacheMiss | 410,381 | 410,442 | 61 | 0.01%
AST.NumASTBytesAllocated | 65,989,585,250 | 66,112,063,336 | 122,478,085 | 0.19%
AST.NumASTScopeExpansions | 9,610,904 | 9,611,733 | 829 | 0.01%
AST.NumASTScopeLookups | 10,533,861 | 10,535,264 | 1,402 | 0.01%
AST.NumDecls | 357,853 | 357,971 | 118 | 0.03%
AST.NumDependencies | 269,394 | 269,386 | -8 | -0.0%
AST.NumIncrementalDependencies | 28,792 | 28,792 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 322,193 | 322,191 | -2 | -0.0%
AST.NumLocalTypeDecls | 6,812 | 6,812 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,409 | 6,409 | 0 | 0.0%
AST.NumModuleLookupValue | 118,493,568 | 118,503,963 | 10,394 | 0.01%
AST.NumObjCMethods | 95,288 | 95,299 | 11 | 0.01%
AST.NumOperators | 885 | 890 | 5 | 0.62%
AST.NumReferencedDynamicNames | 148 | 148 | 0 | 0.0%
AST.NumReferencedMemberNames | 15,779,877 | 15,779,972 | 95 | 0.0%
AST.NumReferencedTopLevelNames | 1,402,981 | 1,402,944 | -37 | -0.0%
AST.NumSourceBuffers | 408,394 | 408,389 | -5 | -0.0%
AST.NumSourceLines | 13,476,934 | 13,482,124 | 5,190 | 0.04%
AST.NumSourceLinesPerSecond | 5,086,838 | 5,058,338 | -28,500 | -0.56%
AST.NumTotalClangImportedEntities | 1,798,546 | 1,796,856 | -1,690 | -0.09%
Frontend.MaxMallocUsage | 931,142,638,104 | 931,541,544,648 | 398,906,544 | 0.04%
Frontend.NumInstructionsExecuted | 144,966,420,101,645 | 145,119,598,291,718 | 153,178,190,073 | 0.11%
Frontend.NumProcessFailures | 40 | 40 | 0 | 0.0%
IRGen.IRGenRequest | 29,613 | 29,613 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 331,400 | 331,400 | 0 | 0.0%
IRModule.NumIRAliases | 268,504 | 268,504 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 8,901,743 | 8,901,743 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,855,427 | 4,855,427 | 0 | 0.0%
IRModule.NumIRGlobals | 4,549,292 | 4,549,292 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 94,129,297 | 94,129,297 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 119,972 | 119,972 | 0 | 0.0%
IRModule.NumIRValueSymbols | 9,494,900 | 9,494,900 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 2,655,829,332 | 2,655,894,712 | 65,380 | 0.0%
Parse.ExportedSourceFileRequest | 4 | 4 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 680,189 | 680,023 | -166 | -0.02%
Parse.NumIterableDeclContextParsed | 871,578 | 871,685 | 106 | 0.01%
Parse.ParseAbstractFunctionBodyRequest | 784,397 | 784,560 | 163 | 0.02%
Parse.ParseMembersRequest | 675,238 | 675,405 | 167 | 0.02%
Parse.ParseSourceFileRequest | 374,442 | 374,441 | -1 | -0.0%
Parse.ParseTopLevelDeclsRequest | 374,442 | 374,441 | -1 | -0.0%
SILGen.ASTLoweringRequest | 30,829 | 30,829 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,804,529 | 3,804,529 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 132,960 | 132,960 | 0 | 0.0%
SILModule.NumSILGenVtables | 40,679 | 40,679 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 235,765 | 235,765 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,244,748 | 4,244,748 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 135,031 | 135,031 | 0 | 0.0%
SILModule.NumSILOptVtables | 41,266 | 41,266 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 274,277 | 274,277 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 151,741 | 151,741 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 330,742 | 330,762 | 20 | 0.01%
Sema.AbstractGenericSignatureRequest | 118,707 | 118,711 | 4 | 0.0%
Sema.AccessLevelRequest | 21,105,075 | 21,272,595 | 167,519 | 0.79%
Sema.ActorIsolationRequest | 45,236,460 | 45,548,741 | 312,280 | 0.69%
Sema.AllMembersRequest | 186,397 | 186,397 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 217 | 217 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,814,021 | 4,816,333 | 2,312 | 0.05%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 51,822 | 51,828 | 6 | 0.01%
Sema.AssociatedConformanceRequest | 187,070 | 187,090 | 19 | 0.01%
Sema.AttachedPropertyWrapperTypeRequest | 35,322 | 35,337 | 15 | 0.04%
Sema.AttachedPropertyWrappersRequest | 31,302,449 | 31,470,891 | 168,441 | 0.54%
Sema.AttachedResultBuilderRequest | 2,074,295 | 2,074,773 | 478 | 0.02%
Sema.BodyInitKindRequest | 85,638 | 85,641 | 2 | 0.0%
Sema.BreakTargetRequest | 7,370 | 7,375 | 4 | 0.06%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 509,949 | 509,949 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 59,797 | 59,821 | 24 | 0.04%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,635 | 10,635 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 59,797 | 59,821 | 24 | 0.04%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,635 | 10,635 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 2,577,764 | 2,578,551 | 786 | 0.03%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 17,757 | 17,740 | -17 | -0.1%
Sema.ClangRecordMemberLookup | 17,757 | 17,740 | -17 | -0.1%
Sema.ClassAncestryFlagsRequest | 153,873 | 153,954 | 81 | 0.05%
Sema.ClosureEffectsRequest | 160,446 | 160,423 | -23 | -0.01%
Sema.ClosureHasExplicitResultRequest | 61,256 | 61,243 | -13 | -0.02%
Sema.CollectOverriddenDeclsRequest | 16,878,968 | 16,975,614 | 96,646 | 0.57%
Sema.CompareDeclSpecializationRequest | 917,230 | 918,229 | 999 | 0.11%
Sema.CompilerPluginLoadRequest | 4 | 4 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 955,793 | 957,233 | 1,440 | 0.15%
Sema.ConformanceHasEffectRequest | 81 | 81 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,467 | 2,467 | 0 | -0.02%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 27,604 | 27,646 | 42 | 0.15%
Sema.CustomAttrTypeRequest | 4,533 | 4,534 | 1 | 0.02%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 64,615 | 64,630 | 14 | 0.02%
Sema.DefaultAndMaxAccessLevelRequest | 129,118 | 129,126 | 8 | 0.01%
Sema.DefaultArgumentExprRequest | 140,183 | 140,260 | 77 | 0.05%
Sema.DefaultArgumentInitContextRequest | 935 | 935 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 17,541 | 17,542 | 1 | 0.01%
Sema.DefaultDefinitionTypeRequest | 2,189 | 2,188 | -1 | -0.07%
Sema.DefaultInitializerIsolation | 1,399,226 | 1,399,727 | 500 | 0.04%
Sema.DefaultTypeRequest | 719,700 | 719,733 | 32 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 92,000,626 | 92,051,950 | 51,324 | 0.06%
Sema.DirectOperatorLookupRequest | 2,513,002 | 2,513,170 | 168 | 0.01%
Sema.DirectPrecedenceGroupLookupRequest | 934,788 | 934,932 | 144 | 0.02%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 42,237,360 | 42,533,042 | 295,682 | 0.7%
Sema.EnumElementExprPatternRequest | 23 | 23 | 0 | 0.0%
Sema.EnumRawTypeRequest | 44,942 | 44,956 | 14 | 0.03%
Sema.EnumRawValuesRequest | 26,478 | 26,494 | 16 | 0.06%
Sema.ExistentialConformsToSelfRequest | 34,910 | 34,931 | 20 | 0.06%
Sema.ExpandAccessorMacros | 1,142,968 | 1,143,116 | 147 | 0.01%
Sema.ExpandBodyMacroRequest | 480,767 | 480,667 | -100 | -0.02%
Sema.ExpandChildTypeRefinementContextsRequest | 794,793 | 794,956 | 163 | 0.02%
Sema.ExpandExtensionMacros | 1,128,577 | 1,132,091 | 3,514 | 0.31%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 75,293,521 | 75,645,296 | 351,774 | 0.47%
Sema.ExpandPeerMacroRequest | 54,543,606 | 54,892,289 | 348,683 | 0.64%
Sema.ExpandPreambleMacroRequest | 480,767 | 480,667 | -100 | -0.02%
Sema.ExpandSynthesizedMemberMacroRequest | 335,420 | 335,452 | 31 | 0.01%
Sema.ExplicitCaughtTypeRequest | 36,766 | 36,736 | -30 | -0.08%
Sema.ExprPatternMatchRequest | 95,820 | 95,820 | 0 | 0.0%
Sema.ExtendedNominalRequest | 760,160 | 760,166 | 6 | 0.0%
Sema.ExtendedTypeRequest | 242,276 | 242,310 | 34 | 0.01%
Sema.ExternalMacroDefinitionRequest | 4 | 4 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 3,202,409 | 3,203,112 | 703 | 0.02%
Sema.FunctionOperatorRequest | 20,269 | 20,279 | 10 | 0.05%
Sema.GenericParamListRequest | 7,737,392 | 7,746,996 | 9,604 | 0.12%
Sema.GenericSignatureRequest | 3,829,782 | 3,831,807 | 2,025 | 0.05%
Sema.GetDestructorRequest | 47,484 | 47,492 | 7 | 0.02%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 40 | 40 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 21 | 21 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 8 | 8 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 13 | 13 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 142 | 142 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 49,140,166 | 49,472,510 | 332,344 | 0.68%
Sema.GlobalActorInstanceRequest | 2,565 | 2,582 | 16 | 0.64%
Sema.HasCircularInheritedProtocolsRequest | 19,176 | 19,164 | -12 | -0.07%
Sema.HasCircularRawValueRequest | 27,465 | 27,483 | 17 | 0.06%
Sema.HasDefaultInitRequest | 117,660 | 117,698 | 37 | 0.03%
Sema.HasDynamicCallableAttributeRequest | 32,996 | 32,927 | -69 | -0.21%
Sema.HasDynamicMemberLookupAttributeRequest | 603,938 | 604,516 | 577 | 0.1%
Sema.HasImportsMatchingFlagRequest | 740,821 | 740,818 | -3 | -0.0%
Sema.HasInitAccessorRequest | 6,065,124 | 6,081,683 | 16,559 | 0.27%
Sema.HasIsolatedSelfRequest | 54,781,695 | 55,135,785 | 354,090 | 0.65%
Sema.HasMemberwiseInitRequest | 62,689 | 62,714 | 24 | 0.04%
Sema.HasMissingDesignatedInitializersRequest | 33,411 | 33,413 | 2 | 0.01%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 17,046 | 17,054 | 8 | 0.05%
Sema.HasStorageRequest | 17,550,960 | 17,700,292 | 149,331 | 0.85%
Sema.HasUserDefinedDesignatedInitRequest | 117,670 | 117,708 | 37 | 0.03%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.ImplicitKnownProtocolConformanceRequest | 278,430 | 279,008 | 578 | 0.21%
Sema.InferredGenericSignatureRequest | 255,927 | 255,905 | -21 | -0.01%
Sema.InheritedDeclsReferencedRequest | 8,272,825 | 8,281,926 | 9,101 | 0.11%
Sema.InheritedProtocolsRequest | 873,432 | 874,759 | 1,327 | 0.15%
Sema.InheritedTypeRequest | 510,460 | 510,329 | -130 | -0.03%
Sema.InheritsSuperclassInitializersRequest | 42,293 | 42,301 | 8 | 0.02%
Sema.InitAccessorPropertiesRequest | 82,127 | 82,133 | 6 | 0.01%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 225,036 | 225,080 | 44 | 0.02%
Sema.InterfaceTypeRequest | 26,282,825 | 26,383,551 | 100,726 | 0.38%
Sema.InvertibleAnnotationRequest | 1,135,193 | 1,135,114 | -79 | -0.01%
Sema.IsABICompatibleOverrideRequest | 385,929 | 385,929 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 672,866 | 673,054 | 187 | 0.03%
Sema.IsActorRequest | 2,697,348 | 2,703,667 | 6,318 | 0.23%
Sema.IsCCompatibleFuncDeclRequest | 0 | 0 | 0 | 0.0%
Sema.IsCallAsFunctionNominalRequest | 7,282 | 7,277 | -5 | -0.07%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 73,017 | 72,984 | -32 | -0.04%
Sema.IsDefaultActorRequest | 52,894 | 52,894 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,905,373 | 2,911,857 | 6,484 | 0.22%
Sema.IsDynamicRequest | 40,955,860 | 41,268,182 | 312,322 | 0.76%
Sema.IsEscapableRequest | 0 | 0 | 0 | 0.0%
Sema.IsFinalRequest | 40,880,322 | 41,191,699 | 311,377 | 0.76%
Sema.IsFunctionBodySkippedRequest | 2,075,547 | 2,076,135 | 587 | 0.03%
Sema.IsGetterMutatingRequest | 916,220 | 916,430 | 210 | 0.02%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,422,264 | 4,421,962 | -302 | -0.01%
Sema.IsNonUserModuleRequest | 106,779 | 106,797 | 18 | 0.02%
Sema.IsNoncopyableRequest | 0 | 0 | 0 | 0.0%
Sema.IsObjCRequest | 3,389,024 | 3,391,049 | 2,025 | 0.06%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 822,108 | 822,271 | 163 | 0.02%
Sema.IsSingleValueStmtRequest | 407,240 | 407,184 | -55 | -0.01%
Sema.IsStaticRequest | 1,841,102 | 1,841,513 | 411 | 0.02%
Sema.LazyStoragePropertyRequest | 2,597 | 2,597 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 633,898 | 633,898 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 3,501,980 | 3,525,411 | 23,431 | 0.67%
Sema.LookupConformanceInModuleRequest | 74,553,501 | 74,799,573 | 246,072 | 0.33%
Sema.LookupInModuleRequest | 10,544,777 | 10,544,678 | -99 | -0.0%
Sema.LookupInfixOperatorRequest | 84,411 | 84,409 | -2 | -0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 34,234 | 34,240 | 6 | 0.02%
Sema.LookupPrefixOperatorRequest | 372 | 374 | 2 | 0.54%
Sema.MacroDefinitionRequest | 8 | 8 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,208 | 3,208 | 0 | 0.0%
Sema.MemberwiseInitPropertiesRequest | 82,129 | 82,135 | 6 | 0.01%
Sema.ModuleImplicitImportsRequest | 10,635 | 10,635 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 3,911,882 | 3,909,864 | -2,018 | -0.05%
Sema.NamedLazyMemberLoadSuccessCount | 67,660,879 | 67,743,174 | 82,295 | 0.12%
Sema.NamingPatternRequest | 270,584 | 270,639 | 54 | 0.02%
Sema.NeedsNewVTableEntryRequest | 1,237,889 | 1,237,889 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 380,227 | 380,396 | 168 | 0.04%
Sema.NumAccessorsSynthesized | 616,949 | 617,273 | 324 | 0.05%
Sema.NumConformancePathsRecorded | 968,195 | 968,823 | 628 | 0.06%
Sema.NumConformancesDeserialized | 6,540,360 | 6,555,985 | 15,624 | 0.24%
Sema.NumConstraintScopes | 85,671,364 | 85,679,128 | 7,764 | 0.01%
Sema.NumConstraintsConsideredForEdgeContraction | 2,878,343 | 2,878,311 | -32 | -0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 103,752,954 | 104,191,121 | 438,166 | 0.42%
Sema.NumDeclsTypechecked | 3,645,130 | 3,646,266 | 1,135 | 0.03%
Sema.NumLazyIterableDeclContexts | 15,199,497 | 15,236,904 | 37,407 | 0.25%
Sema.NumLazyRequirementSignatures | 1,156,617 | 1,158,194 | 1,577 | 0.14%
Sema.NumLazyRequirementSignaturesLoaded | 809,515 | 811,003 | 1,488 | 0.18%
Sema.NumLeafScopes | 67,107,361 | 67,114,681 | 7,320 | 0.01%
Sema.NumRequirementMachineCompletionSteps | 27,793,844 | 27,883,961 | 90,117 | 0.32%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,425,362 | 1,432,185 | 6,823 | 0.48%
Sema.NumRequirementMachines | 2,408,277 | 2,415,332 | 7,055 | 0.29%
Sema.NumTypesDeserialized | 27,481,674 | 27,569,058 | 87,384 | 0.32%
Sema.NumUnloadedLazyIterableDeclContexts | 11,391,699 | 11,400,312 | 8,613 | 0.08%
Sema.ObjCInterfaceAndImplementationRequest | 2,594,834 | 2,595,371 | 537 | 0.02%
Sema.OpaqueReadOwnershipRequest | 560,867 | 561,090 | 223 | 0.04%
Sema.OpaqueResultTypeRequest | 961 | 962 | 1 | 0.16%
Sema.OperatorPrecedenceGroupRequest | 943 | 946 | 3 | 0.37%
Sema.OverriddenDeclsRequest | 25,658,244 | 25,838,278 | 180,033 | 0.7%
Sema.ParamSpecifierRequest | 2,364,696 | 2,364,942 | 246 | 0.01%
Sema.PatternBindingCheckedAndContextualizedInitRequest | 447,341 | 447,386 | 45 | 0.01%
Sema.PatternBindingEntryRequest | 900,704 | 900,884 | 180 | 0.02%
Sema.PatternTypeRequest | 1,071,556 | 1,071,723 | 166 | 0.02%
Sema.PolymorphicEffectKindRequest | 1,180,068 | 1,179,888 | -180 | -0.02%
Sema.PolymorphicEffectRequirementsRequest | 92 | 92 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 7,650,577 | 7,656,188 | 5,611 | 0.07%
Sema.PreCheckFunctionBodyRequest | 474,808 | 474,733 | -75 | -0.02%
Sema.PreCheckResultBuilderRequest | 1,502 | 1,502 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 649,677 | 649,583 | -93 | -0.01%
Sema.PrimaryAssociatedTypesRequest | 21,277 | 21,465 | 188 | 0.88%
Sema.PrimarySourceFilesRequest | 10,635 | 10,635 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 3,116,854 | 3,117,018 | 163 | 0.01%
Sema.PropertyWrapperBackingPropertyTypeRequest | 35,322 | 35,337 | 15 | 0.04%
Sema.PropertyWrapperInitializerInfoRequest | 795,783 | 795,946 | 163 | 0.02%
Sema.PropertyWrapperLValuenessRequest | 2,718 | 2,718 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 909,010 | 909,218 | 208 | 0.02%
Sema.PropertyWrapperTypeInfoRequest | 713 | 714 | 1 | 0.14%
Sema.ProtocolDependenciesRequest | 695,107 | 696,146 | 1,039 | 0.15%
Sema.ProtocolRequirementsRequest | 84,907 | 84,919 | 12 | 0.01%
Sema.ProtocolRequiresClassRequest | 50,261 | 50,136 | -125 | -0.25%
Sema.ProvideDefaultImplForRequest | 16,878,968 | 16,975,614 | 96,646 | 0.57%
Sema.QualifiedLookupRequest | 11,271,313 | 11,283,311 | 11,998 | 0.11%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,381,391 | 1,381,391 | 0 | 0.0%
…